Internationalize pay-in sankey values
